Component state registry and access implementation (#587)
