+++ inventory-framework-platform/src/main/java/me/devnatan/inventoryframework/state/StateAccessImpl.java
@@ -5,13 +5,10 @@ import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import me.devnatan.inventoryframework.VirtualView;
-import me.devnatan.inventoryframework.component.ItemComponentBuilder;
 import me.devnatan.inventoryframework.component.Pagination;
 import me.devnatan.inventoryframework.component.PaginationBuilder;
 import me.devnatan.inventoryframework.component.PaginationImpl;
 import me.devnatan.inventoryframework.component.PaginationValueConsumer;
-import me.devnatan.inventoryframework.context.IFContext;
-import me.devnatan.inventoryframework.internal.ElementFactory;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
@@ -20,16 +17,13 @@ import org.jetbrains.annotations.NotNull;
  * this library. No compatibility guarantees are provided. </i></b>
  */
 @ApiStatus.Internal
-public final class StateAccessImpl<CONTEXT extends IFContext, ITEM_BUILDER extends ItemComponentBuilder>
-        implements StateAccess<CONTEXT, ITEM_BUILDER> {
+public final class StateAccessImpl<CONTEXT, ITEM_BUILDER> implements StateAccess<CONTEXT, ITEM_BUILDER> {
 
     private final VirtualView caller;
-    private final ElementFactory elementFactory;
     private final StateRegistry stateRegistry;
 
-    public StateAccessImpl(VirtualView caller, ElementFactory elementFactory, StateRegistry stateRegistry) {
+    public StateAccessImpl(VirtualView caller, StateRegistry stateRegistry) {
         this.caller = caller;
-        this.elementFactory = elementFactory;
         this.stateRegistry = stateRegistry;
     }
 
@@ -172,37 +166,37 @@ public final class StateAccessImpl<CONTEXT extends IFContext, ITEM_BUILDER exten
     @Override
     public <T> PaginationBuilder<CONTEXT, ITEM_BUILDER, T> buildPaginationState(
             @NotNull List<? super T> sourceProvider) {
-        return new PaginationBuilder<>(this.elementFactory, sourceProvider, false, false);
+        return new PaginationBuilder<>(sourceProvider, false, false);
     }
 
     @Override
     public <T> PaginationBuilder<CONTEXT, ITEM_BUILDER, T> buildComputedPaginationState(
             @NotNull Function<CONTEXT, List<? super T>> sourceProvider) {
-        return new PaginationBuilder<>(this.elementFactory, sourceProvider, false, true);
+        return new PaginationBuilder<>(sourceProvider, false, true);
     }
 
     @Override
     public <T> PaginationBuilder<CONTEXT, ITEM_BUILDER, T> buildComputedAsyncPaginationState(
             @NotNull Function<CONTEXT, CompletableFuture<List<T>>> sourceProvider) {
-        return new PaginationBuilder<>(this.elementFactory, sourceProvider, true, true);
+        return new PaginationBuilder<>(sourceProvider, true, true);
     }
 
     @Override
     public <T> PaginationBuilder<CONTEXT, ITEM_BUILDER, T> buildLazyPaginationState(
             @NotNull Supplier<List<? super T>> sourceProvider) {
-        return new PaginationBuilder<>(this.elementFactory, sourceProvider, false, false);
+        return new PaginationBuilder<>(sourceProvider, false, false);
     }
 
     @Override
     public <T> PaginationBuilder<CONTEXT, ITEM_BUILDER, T> buildLazyPaginationState(
             @NotNull Function<CONTEXT, List<? super T>> sourceProvider) {
-        return new PaginationBuilder<>(this.elementFactory, sourceProvider, false, false);
+        return new PaginationBuilder<>(sourceProvider, false, false);
     }
 
     @Override
     public <T> PaginationBuilder<CONTEXT, ITEM_BUILDER, T> buildLazyAsyncPaginationState(
             @NotNull Function<CONTEXT, CompletableFuture<List<T>>> sourceProvider) {
-        return new PaginationBuilder<>(this.elementFactory, sourceProvider, true, false);
+        return new PaginationBuilder<>(sourceProvider, true, false);
     }
 
     <V> State<Pagination> createPaginationState(@NotNull PaginationBuilder<CONTEXT, ITEM_BUILDER, V> builder) {
